Skip generating files if they already exist.

+++ lib/vedeu/cli/generator/helpers.rb
@@ -27,18 +27,35 @@ module Vedeu
       # @param destination [String]
       # @return [void]
       def copy_file(source, destination)
-        Vedeu.log_stdout(type: :create, message: "#{destination}")
+        if File.exist?(destination)
+          skipped_file(destination)
 
-        FileUtils.cp(source, destination)
+        else
+          Vedeu.log_stdout(type: :create, message: "#{destination}")
+
+          FileUtils.cp(source, destination)
+        end
       end
 
       # @param source [String]
       # @param destination [String]
       # @return [void]
       def make_file(source, destination)
-        Vedeu.log_stdout(type: :create, message: "#{destination}")
+        if File.exist?(destination)
+          skipped_file(destination)
 
-        File.write(destination, parse(source))
+        else
+          Vedeu.log_stdout(type: :create, message: "#{destination}")
+
+          File.write(destination, parse(source))
+        end
+      end
+
+      # @param destination [String]
+      # @return [void]
+      def skipped_file(destination)
+        Vedeu.log_stdout(type:    :create,
+                         message: "#{destination} exists, skipped.")
       end
 
       # @param destination [String]
